old/requirements.txt
pyautogui==0.9.54
openai==0.28.1
python-dotenv==1.0.0